+++ rust/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.3.0] - 2026-04-14
+
+### Added
+
+* X25519 interface - this is only intended to support rustTLS
+
 ## [0.2.2] - 2025-10-18
 
 ### Added